Expose disjuncts produced during loss conversion
